Don't cancel token before stop session manager
I don't know if the order here was intended, but I had some problems in a project.
There was a test which stopped the broker and started it again after `StopAsync` was awaited.
While stopping the server there was the following exception (which lead to a `SocketException` while starting again -> `System.Net.Sockets.SocketException (10048): Only one usage of each socket address (protocol/network address/port) is normally permitted`)
```
   at System.Collections.Concurrent.BlockingCollection`1.TryAddWithNoTimeValidation(T item, Int32 millisecondsTimeout, CancellationToken cancellationToken)
   at MQTTnet.Server.MqttClientSessionsManager.EnqueueApplicationMessage(MqttClientSession senderClientSession, MqttPublishPacket publishPacket)
   at MQTTnet.Server.MqttClientSession.Stop(MqttClientDisconnectType type, Boolean isInsideSession)
   at MQTTnet.Server.MqttClientSessionsManager.Stop()
   at MQTTnet.Server.MqttServer.StopAsync()
```
So the `EnqueueApplicationMessage` from `MqttClientSessionsManager` was called because the `MqttClientSession` did have a last will message, but since the cancellation token was already canceled, this exception was thrown. Therefore the broker was not closed correctly.